pyproject.toml
[project]
name = "badminton-sense"
version = "0.1.0"
description = "Deep learning stroke classification from monocular badminton video via spatio-temporal pose estimation"
requires-python = ">=3.10"
authors = [
    {name = "Abhay Singh Thakur", email = "thakur22429s@gmail.com"}
]

dependencies = [
    "torch>=2.0",
    "torchvision>=0.15",
    "mediapipe>=0.10",
    "opencv-python>=4.8",
    "numpy>=1.24",
    "pandas>=2.0",
    "scikit-learn>=1.3",
    "matplotlib>=3.7",
    "seaborn>=0.12",
    "streamlit>=1.28",
    "pyyaml>=6.0",
    "tqdm>=4.65",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ipykernel>=6.0",
    "jupyter>=1.0",
]
tracking = [
    "wandb>=0.15",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
